+++ src/main/java/com/kul/Estypendia/repository/PaymentsLogRepo.java
@@ -4,9 +4,10 @@ import com.kul.Estypendia.model.PaymentsLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 
 @Repository
 public interface PaymentsLogRepo  extends JpaRepository<PaymentsLog, Integer> {
-    List<PaymentsLog> findAllByStudentId (Integer studentId);
+    List<PaymentsLog> findAllByStudentId (BigInteger studentId);
 }
